pyproject.toml
[project]
name = "quant-qa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "faiss-cpu>=1.13.0",
    "pymupdf>=1.26.6",
    "sentence-transformers>=5.1.2",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]
